+++ tests/service/privacy_request/request_runner_service_test.py
@@ -13,7 +13,7 @@ from sqlalchemy.orm import Session
 from fidesops.common_exceptions import ClientUnsuccessfulException, PrivacyRequestPaused
 from fidesops.core.config import config
 from fidesops.db.session import get_db_session
-from fidesops.models.policy import PolicyPostWebhook
+from fidesops.models.policy import PausedStep, PolicyPostWebhook
 from fidesops.models.privacy_request import (
     ActionType,
     ExecutionLog,
@@ -80,16 +80,27 @@ def test_start_processing_doesnt_overwrite_started_processing_at(
 
 
 @mock.patch(
-    "fidesops.service.privacy_request.request_runner_service.PrivacyRequestRunner.run_webhooks_and_report_status"
+    "fidesops.service.privacy_request.request_runner_service.PrivacyRequestRunner.run_webhooks_and_report_status",
 )
+@mock.patch(
+    "fidesops.service.privacy_request.request_runner_service.run_access_request"
+)
+@mock.patch("fidesops.service.privacy_request.request_runner_service.run_erasure")
 def test_from_graph_resume_does_not_run_pre_webhooks(
+    run_erasure,
+    run_access,
     run_webhooks,
     db: Session,
     privacy_request: PrivacyRequest,
     privacy_request_runner: PrivacyRequestRunner,
+    erasure_policy,
 ) -> None:
     privacy_request.started_processing_at = None
-    wait_for(privacy_request_runner.submit(from_step=ActionType.access))
+    privacy_request.policy = erasure_policy
+    privacy_request.save(db)
+
+    privacy_request.started_processing_at = None
+    wait_for(privacy_request_runner.submit(from_step=PausedStep.access))
 
     _sessionmaker = get_db_session()
     db = _sessionmaker()
@@ -100,6 +111,45 @@ def test_from_graph_resume_does_not_run_pre_webhooks(
     # Starting privacy request in the middle of the graph means we don't run pre-webhooks again
     assert run_webhooks.call_count == 1
     assert run_webhooks.call_args[1]["webhook_cls"] == PolicyPostWebhook
+
+    assert run_access.call_count == 1  # Access request runs
+    assert run_erasure.call_count == 1  # Erasure request runs
+
+
+@mock.patch(
+    "fidesops.service.privacy_request.request_runner_service.PrivacyRequestRunner.run_webhooks_and_report_status",
+)
+@mock.patch(
+    "fidesops.service.privacy_request.request_runner_service.run_access_request"
+)
+@mock.patch("fidesops.service.privacy_request.request_runner_service.run_erasure")
+def test_resume_privacy_request_from_erasure(
+    run_erasure,
+    run_access,
+    run_webhooks,
+    db: Session,
+    privacy_request: PrivacyRequest,
+    privacy_request_runner: PrivacyRequestRunner,
+    erasure_policy,
+) -> None:
+    privacy_request.started_processing_at = None
+    privacy_request.policy = erasure_policy
+    privacy_request.save(db)
+
+    wait_for(privacy_request_runner.submit(from_step=PausedStep.erasure))
+
+    _sessionmaker = get_db_session()
+    db = _sessionmaker()
+
+    privacy_request = PrivacyRequest.get(db=db, id=privacy_request.id)
+    assert privacy_request.started_processing_at is not None
+
+    # Starting privacy request in the middle of the graph means we don't run pre-webhooks again
+    assert run_webhooks.call_count == 1
+    assert run_webhooks.call_args[1]["webhook_cls"] == PolicyPostWebhook
+
+    assert run_access.call_count == 0  # Access request skipped
+    assert run_erasure.call_count == 1  # Erasure request runs
 
 
 def get_privacy_request_results(